Fixed the various issues
Should be working now! Changed the bread smite to a 'generic item' smite and then made it a specifcally breaded smite.

+++ code/modules/mob/living/living.dm
@@ -1190,18 +1190,22 @@
 	return FALSE
 
 /mob/living/proc/plushify(plushie_overide, curse_time = 10 MINUTES)
-	var/mob/living/simple_animal/shade/sword/bread/plushvictim = new(get_turf(src))
+	var/mob/living/simple_animal/shade/sword/generic_item/plushvictim = new(get_turf(src))
 	var/obj/item/toy/plushie/plush_type = pick(subtypesof(/obj/item/toy/plushie) - typesof(/obj/item/toy/plushie/fluff) - typesof(/obj/item/toy/plushie/carpplushie)) //exclude the base type.
 	if(plushie_overide)
 		plush_type = plushie_overide
 	var/obj/item/toy/plushie/plush_outcome = new plush_type(get_turf(src))
 	plushvictim.forceMove(plush_outcome)
 	plushvictim.key = key
-	plushvictim.RegisterSignal(plush_outcome, COMSIG_PARENT_QDELETING, TYPE_PROC_REF(/mob/living/simple_animal/shade/sword/bread, handle_bread_deletion))
+	plushvictim.RegisterSignal(plush_outcome, COMSIG_PARENT_QDELETING, TYPE_PROC_REF(/mob/living/simple_animal/shade/sword/generic_item, handle_item_deletion))
 	plushvictim.name = name
 	plush_outcome.name = "[name] plushie"
-	if(istype(src, /mob/living/silicon/robot))
-		var/mob/living/silicon/robot/R = src
-		QDEL_NULL(R.mmi)
-	qdel(src)
+	if(curse_time == -1)
+		qdel(src)
+	else
+		plush_outcome.cursed_plushie_victim = src
+		forceMove(plush_outcome)
+		notransform = TRUE
+		status_flags |= GODMODE
+		addtimer(CALLBACK(plush_outcome, TYPE_PROC_REF(/obj/item/toy/plushie, un_plushify)), curse_time)
 	to_chat(plushvictim, "<span class='warning'>You have been cursed into an enchanted plush doll! At least you can still move around a bit...</span>")
